Clean page components (containers) up, simplify DOM
